requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==2.1.1
Django==4.1.4
django-cors-headers==3.13.0
django-environ==0.9.0
djangorestframework==3.14.0
fedex==2.4.1
idna==3.4
mysqlclient==2.1.1
pytz==2022.7
requests==2.28.1
sqlparse==0.4.3
suds-jurko==0.6
urllib3==1.26.13
